Change usize to u64 for all variables that aren't indices into byte-based structures (#34)
This changes all types that index bits instead of words to u64, adjusts the necessary conversions, and suppresses truncation warnings where truncations are impossible. 
Furthermore, it also modifies the usage, documentation, and implementation of iterator types, because various iterator functions are limited to `usize`. We avoid truncation where possible and panic where it cannot be avoided (which only applies to len() and count())

+++ benches/select_iter.rs
@@ -15,11 +15,11 @@ fn bench_select_iter(b: &mut Criterion) {
         group.bench_with_input(BenchmarkId::new("select queries", l), &l, |b, _| {
             b.iter_custom(|iters| {
                 let mut time = Duration::new(0, 0);
-                let mut i = 0usize;
+                let mut i = 0;
                 let rank1 = bit_vec.rank1(bit_vec.len());
 
                 let start = Instant::now();
-                while (i as u64) < iters {
+                while (i) < iters {
                     black_box(bit_vec.select1(i % rank1));
                     i += 1;
                 }
